% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_profile.R
\name{aggregate_features}
\alias{aggregate_features}
\title{Aggregate feature descriptors}
\usage{
aggregate_features(plate.ids, plate.dir, write.dir)
}
\arguments{
\item{plate.dir}{(character) path to plate directory, containing the plate
directory with image features for the selected plate.}

\item{plates}{(character) identifiers of the plates whose ks profiles will be
loaded and aggregated.}

\item{write.path}{(character) path to write aggregated ks profiles to.}
}
\value{
None. Will save a csv file containing KS profiles for all selected
  plates.
}
\description{
Function to load in and aggregate profiles generated for multiple plates.
}
