% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ksTest}
\alias{ksTest}
\title{Signed KS test
This functions is a modified version of the R-base function ks.test. It
computes a signed KS statistic measuring the maximum difference between a
distribution x and some reference distribution y.}
\usage{
ksTest(x, y)
}
\arguments{
\item{x}{(numeric) data from distribution 1}

\item{y}{(numeric) data from distribution 2}
}
\value{
Signed KS statistic comparing distributions of x and y
}
\description{
Signed KS test
This functions is a modified version of the R-base function ks.test. It
computes a signed KS statistic measuring the maximum difference between a
distribution x and some reference distribution y.
}
