% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadControl}
\alias{loadControl}
\title{Load control data}
\usage{
loadControl(
  xmeta,
  plate.dir,
  type = "cbfeature",
  control.cl = NULL,
  control.cpd = NULL,
  control.usg = NULL
)
}
\description{
Load control data
}
