% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_feature.R
\name{loadFeature}
\alias{loadFeature}
\title{Load feature descriptors}
\usage{
loadFeature(plate)
}
\arguments{
\item{plate}{(character) path to plate directory}
}
\value{
data.frame indicating feature name, channel, and category
}
\description{
Core function for generating a table of feature descriptors for a select
plate.
}
