% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metadata.R
\name{loadMeta}
\alias{loadMeta}
\title{Load metadata}
\usage{
loadMeta(meta.file)
}
\arguments{
\item{meta.file}{(character): path to file containing plate metadata.}
}
\value{
a data table with one entry per well, indicating metadata associated
  with each well.
}
\description{
Load in metadata, stored as excel spreadsheet with standard formatting,  
for a select plate. TODO: Note on standard formatting.
}
