% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_profile.R
\name{wellKS}
\alias{wellKS}
\title{Evaluate KS given well level and control matrices}
\usage{
wellKS(xwell, xcontrol, id.feat = NULL)
}
\description{
Evaluate KS given well level and control matrices
}
