% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_profile.R
\name{generateProfile}
\alias{generateProfile}
\title{Generate KS profile}
\usage{
generateProfile(
  plate,
  xmeta,
  plate.dir,
  control.variable,
  controls,
  type = "cbfeature",
  n.bs = 0,
  n.core = 1
)
}
\arguments{
\item{plate}{(character) identifier for plate to generate KS profiles for.
Must match a plateID in xmeta.}

\item{xmeta}{(data.table) well-level metadata as returned by loadMeta.}

\item{plate.dir}{(character) path to plate directory, containing the plate
directory with image features for the selected plate.}

\item{type}{(character) type of features to use for KS profile generation
(one of: cbfeature, operetta)}

\item{n.core}{(int) number of cores to parallelize over.}

\item{control.cl}{(character) cell lines to use as KS reference population.}

\item{control.cpd}{(character) compound to use as KS reference
population.}

\item{control.usg}{(character) cell line usage to set as KS reference
population.}
}
\value{
None. Writes KS profiles to plate directory.
}
\description{
Core function for generating KS profile for all wells on a plate.
}
